pyproject.toml
[tool.poetry]
name = "MetaMorph"
version = "1.0.0"
description = "AI Assistance to Transform Lo-Fi Sketches to Higher Fidelities"
authors = ["Vinoth Pandian <vinothpandian@live.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
tensorflow = "1.9.0"
fastapi = "^0.61.2"
uvicorn = {extras = ["standard"], version = "^0.12.0"}
aiofiles = "^0.6.0"
python-multipart = "^0.0.5"
opencv-python = "^4.4.0"
typing-extensions = "^3.7.4"
pandas = "0.23.4"
imutils = "^0.5.3"
pillow = "^8.0.1"
contextlib2 = "^0.6.0"
Pebble = "^4.5.3"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
rope = "^0.18.0"
black = "^20.8b1"


[tool.black]
line-length = 100

[tool.pylint.MASTER]
extension-pkg-whitelist = ["cv2"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


